% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{as.catmaidmesh}
\alias{as.catmaidmesh}
\alias{as.catmaidmesh.default}
\alias{as.catmaidmesh.hxsurf}
\title{Convert a mesh to catmaid format}
\usage{
as.catmaidmesh(x, ...)

\method{as.catmaidmesh}{hxsurf}(x, ...)

\method{as.catmaidmesh}{default}(x, title = NULL, comment = NULL, ...)
}
\arguments{
\item{x}{A mesh object such as hxsurf or a list with raw vertices (Nx3) and 
indices (Nx3).}

\item{...}{Additional fields for the catmaid mesh object}

\item{title}{The title of the object as it will appear in the catmaid volume manager}

\item{comment}{An informative comment - e.g. how the mesh was generated.}
}
\description{
Convert a mesh to catmaid format
}
\references{
\url{https://github.com/catmaid/CATMAID/blob/master/sphinx-doc/source/volumes.rst}
}

