% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{catmaid-package}
\alias{catmaid}
\alias{catmaid-package}
\title{Programmatic access to the CATMAID API}
\description{
See http://catmaid.org/.
}
\details{
Based on https://github.com/acardona/CATMAID/blob/master/scripts/remote/access.py
}
\keyword{package}

