% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{catmaid_add_volume}
\alias{catmaid_add_volume}
\title{Add a 3D mesh surface to catmaid volume manager}
\usage{
catmaid_add_volume(x, conn = NULL, pid = 1, ...)
}
\arguments{
\item{x}{A mesh object such as hxsurf or a list with raw vertices and indices
that will be passed to \code{\link{as.catmaidmesh}}}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{pid}{Project id (default 1)}

\item{...}{Additional arguments passed to \code{\link{as.catmaidmesh}} such
as title, comments etc.}
}
\value{
An invisible list with the post status response on success (the
  function will throw an error on failure)
}
\description{
Add a 3D mesh surface to catmaid volume manager
}
\examples{
\dontrun{
library(elmr)
FAFBNP.surf=xform_brain(JFRC2NP.surf, sample=JFRC2, ref=FAFB14)
for(r in FAFBNP.surf$RegionList){
  comment="Transformed from JFRC2 space onto FAFB14"
  name=paste("v14", sep=".", r)
  catmaid_add_volume(subset(FAFBNP.surf, r), title=name, comment=comment)
}
}
}
\seealso{
\code{\link{as.catmaidmesh}}
}
