% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_adjacency_matrix}
\alias{catmaid_adjacency_matrix}
\title{Get connectivity adjacency matrix between set of neurons}
\usage{
catmaid_adjacency_matrix(
  inputskids,
  outputskids = inputskids,
  pid = 1,
  conn = NULL,
  ...
)
}
\arguments{
\item{inputskids, outputskids}{Input (source) and output (target) skids in any
form understandable to \code{\link{catmaid_skids}}.}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
A matrix, with inputs (sources) as rows and outputs (targets) as
  columns.
}
\description{
Get connectivity adjacency matrix between set of neurons
}
\examples{
da1adj=catmaid_adjacency_matrix("name:DA1")
heatmap(
  da1adj,
  scale = 'none',
  labCol = catmaid_get_neuronnames(colnames(da1adj)),
  labRow = catmaid_get_neuronnames(rownames(da1adj)),
  margins = c(12,12)
)
}
\seealso{
Other connectors: 
\code{\link{catmaid_get_connector_table}()},
\code{\link{catmaid_get_connectors_between}()},
\code{\link{catmaid_get_connectors}()},
\code{\link{catmaid_query_connected}()},
\code{\link{connectors}()}
}
\concept{connectors}
