% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_connection.R
\name{catmaid_connection_setenv}
\alias{catmaid_connection_setenv}
\alias{catmaid_connection_getenv}
\alias{catmaid_connection_unsetenv}
\title{Import/Export catmaid connection details to system variables (e.g. for tests)}
\usage{
catmaid_connection_setenv(conn = NULL, ...)

catmaid_connection_getenv()

catmaid_connection_unsetenv()
}
\arguments{
\item{conn}{A \code{catmaid_connection} object. The default value of NULL
implies that the most recent cached open connection will be used.}

\item{...}{additional arguments passed to \code{catmaid_login}}
}
\value{
\code{catmaid_connection_setenv} returns TRUE or FALSE depending on
  whether variables were set successfully. \code{catmaid_connection_getenv}
  returns a connection object created based on environment variables.
}
\description{
\code{catmaid_connection_setenv} sets environment variables
  based on a \code{catmaid_connection} object.

#' \code{catmaid_connection_getenv} fetches appropriately named
  environment variables and returns them as named character vector.

\code{catmaid_connection_unsetenv} unsets (or removes) the environment
  variables.
}
\details{
\code{catmaid_connection_setenv} will attempt to login if this has
  not already been done
}
\seealso{
\code{\link{catmaid_login}}
}
