% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_stats.R
\name{catmaid_connector_nodes}
\alias{catmaid_connector_nodes}
\title{Get meta information on nodes connected to a CATMAID connector node}
\usage{
catmaid_connector_nodes(
  connector_id,
  node = c("presynaptic", "postsynaptic"),
  pid = 1,
  conn = conn,
  ...
)
}
\arguments{
\item{connector_id}{the ID for a connector of interest}

\item{node}{whether the connector is pre- or post-synaptic to the node of interest}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid_login for details}

\item{...}{methods passed to \code{catmaid_fetch} and \code{catmaid_get_treenode_detail}}
}
\description{
Get information on a CATMAID connector node
}
