% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_annotations.R
\name{catmaid_entities_from_models}
\alias{catmaid_entities_from_models}
\alias{catmaid_models_from_entities}
\title{Convert between entity (neuron) and model (skeleton) ids}
\usage{
catmaid_entities_from_models(skids, pid = 1, conn = NULL, ...)

catmaid_models_from_entities(nids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}

\item{nids}{Neuron (entity) ids}
}
\value{
For \code{catmaid_entities_from_models} an integer vector of \bold{entity ids} each named by the 
  corresponding \bold{model id} (usually a skeleton id).

For \code{catmaid_models_from_entities} a list of \bold{model ids} (usually skeleton ids) named by the 
  corresponding \bold{entity ids}. Formally there may be multiple skeletons per neuron.
}
\description{
\code{catmaid_entities_from_models} converts from neuron ids to skeleton ids

\code{catmaid_models_from_entities} converts skeleton ids to
  from neuron ids
}
\details{
This will normally be used to turn skeleton ids into neuron ids 
  which are used e.g. for annotation purposes. This is probably not something
  that many end users will need but is required e.g. by 
  \code{catmaid_remove_annotations_for_skeletons}.
}
\examples{
ents=catmaid_entities_from_models('ORN PNs left', conn=vfbcatmaid('l1em'))
ents
mods=catmaid_models_from_entities(unname(ents), conn=vfbcatmaid('l1em'))
str(mods)
}
\seealso{
Used by \code{\link{catmaid_remove_annotations_for_skeletons}}
}
