% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catmaid_get_annotationlist}
\alias{catmaid_get_annotationlist}
\title{Get list of annotated neurons (including user information) from CATMAID}
\usage{
catmaid_get_annotationlist(pid, conn = NULL, raw = FALSE, ...)
}
\arguments{
\item{pid}{project and skeleton ids}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
A list containing two data.frames, neurons and users. The users
  data.frame describes the users associated with each neuron, with one row
  for each valid neuron/user pair.
}
\description{
Get list of annotated neurons (including user information) from CATMAID
}
\examples{
\dontrun{
catmaid_get_neuronnames(pid=1)
}
}

