% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catmaid_annotations.R
\name{catmaid_get_annotations_for_skeletons}
\alias{catmaid_get_annotations_for_skeletons}
\title{Get annotations for neurons from CATMAID}
\usage{
catmaid_get_annotations_for_skeletons(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the catmaid_fetch function}
}
\value{
a data.frame containing the following columns \itemize{

  \item skid The skeleton id

  \item annotation The annotation string

  \item id The annotation id

  \item uid The user id for the annotation

  }
}
\description{
Get annotations for neurons from CATMAID
}
\examples{
\dontrun{
catmaid_get_annotations_for_skeletons(skids=c(10418394,4453485))
catmaid_get_annotations_for_skeletons("name:ORN (left|right)")
catmaid_get_annotations_for_skeletons("annotation:ORN PNs$")
}
}
\seealso{
\code{\link{catmaid_fetch}}, \code{\link{catmaid_skids}}
}

