% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_connector_table}
\alias{catmaid_get_connector_table}
\title{Return connector table for a given neuron}
\usage{
catmaid_get_connector_table(skid, direction = c("both", "incoming",
  "outgoing"), pid = 1, conn = NULL, raw = FALSE, ...)
}
\arguments{
\item{skid}{Numeric skeleton id}

\item{direction}{whether to find incoming or outgoing connections}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
A data.frame with columns \itemize{

  \item connector_id

  \item partner_skid

  \item x

  \item y

  \item z

  \item s

  \item confidence

  \item tags

  \item nodes_in_partner

  \item username

  \item partner_treenode_id

  \item last_modified

  }
}
\description{
Return connector table for a given neuron
}
\seealso{
\code{\link{catmaid_get_connectors}}
}

