% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_connectors}
\alias{catmaid_get_connectors}
\title{Return skeleton ids for pre/postsynaptic partners of a set of connector_ids}
\usage{
catmaid_get_connectors(connector_ids, pid = 1, conn = NULL, raw = FALSE,
  ...)
}
\arguments{
\item{connector_ids}{Numeric ids for each connection}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
A data.frame with columns \itemize{

  \item connector_id

  \item pre

  \item post

  }
}
\description{
Return skeleton ids for pre/postsynaptic partners of a set of connector_ids
}

