% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_connectors_between}
\alias{catmaid_get_connectors_between}
\title{Return information about connectors joining sets of pre/postsynaptic skids}
\usage{
catmaid_get_connectors_between(pre_skids, post_skids, pid = 1, conn = NULL,
  raw = FALSE, ...)
}
\arguments{
\item{pre_skids, post_skids}{Skeleton ids in any form understood by
\code{\link{catmaid_skids}}.}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE}) 
or to convert the nodes and connectors lists into processed data.frames 
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\value{
A data.frame with columns \itemize{
  
  \item pre_skid
  
  \item post_skid
  
  \item connector_id
  
  \item pre_node_id
  
  \item post_node_id
  
  \item connector_x
  
  \item connector_y
  
  \item connector_z
  
  \item pre_node_x
  
  \item pre_node_y
  
  \item pre_node_z
  
  \item post_node_x
  
  \item post_node_y
  
  \item post_node_z
  
  \item pre_confidence
  
  \item pre_user
  
  \item post_confidence
  
  \item post_user
  
  }
}
\description{
Return information about connectors joining sets of pre/postsynaptic skids
}
\details{
Each row is a unique set of pre_synaptic node, post_synaptic node, 
  connector_id. A rare (and usually erroneous) scenario is if the same 
  pre_node and post_node are present with two different connector_ids - this 
  would create two rows.
}
\seealso{
Other connectors: \code{\link{catmaid_get_connector_table}},
  \code{\link{catmaid_get_connectors}},
  \code{\link{catmaid_query_connected}},
  \code{\link{connectors}}
}

