% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catmaid_stats.R
\name{catmaid_get_contributor_stats}
\alias{catmaid_get_contributor_stats}
\title{Get contributor statistics for neurons from CATMAID}
\usage{
catmaid_get_contributor_stats(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the catmaid_fetch function}
}
\value{
a list containing different statistics including construction and
  review times (aggregated across all the specified input neurons).
}
\description{
Get contributor statistics for neurons from CATMAID
}
\examples{
\dontrun{
catmaid_get_contributor_stats(skids=c(10418394,4453485))
}
}
\seealso{
\code{\link{catmaid_get_review_status}}
}

