% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_labels.R
\name{catmaid_get_labels}
\alias{catmaid_get_labels}
\title{Get/Set/Remove CATMAID treenode labels (aka tags)}
\usage{
catmaid_get_labels(treenodes = NULL, connectors = NULL, pid = 1,
  conn = NULL, ...)
}
\arguments{
\item{treenodes, connectors}{Integer ids of nodes to be queried}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by 
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection 
object will be generated using the values of the \bold{catmaid.*} package 
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function}
}
\value{
a data.frame with columns \itemize{

  \item id (integer)

  \item type (treenode or connector)

  \item label (the tag)

  }

  A zero row data.frame will be returned if there are no results.
}
\description{
\code{catmaid_get_labels} gets labels (tags) for specified
  nodes.
}
\seealso{
Other labels: \code{\link{catmaid_get_label_stats}}
}
