% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_search.R
\name{catmaid_get_neuronid}
\alias{catmaid_get_neuronid}
\title{Get the CATMAID neuron ID that corresponds to the skeleton ID}
\usage{
catmaid_get_neuronid(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{a vector of skeleton IDs or argument applicable to \code{catmaid_get_neuronid}}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see \code{catmaid::catmaid_login} for details}

\item{...}{methods passed to  \code{catmaid_fetch}}
}
\description{
Retrieve the neuron IDs for given skeleton IDs. This is typically the skeleton ID + 1, and is often, but not always accurately, kept by CATMAID tracers in the name of a neuron.
}
