% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catmaid_get_neuronnames}
\alias{catmaid_get_neuronnames}
\title{Get names of neurons from CATMAID}
\usage{
catmaid_get_neuronnames(pid, skids, ...)
}
\arguments{
\item{pid}{Project id}

\item{skids}{One or more skeleton ids}

\item{...}{Additional arguments passed to the catmaid_GET function}
}
\value{
a named character vector of neuron names. Missing values will be
  represented by a \code{NA_character} value.
}
\description{
Get names of neurons from CATMAID
}
\examples{
\dontrun{
catmaid_get_neuronnames(pid=1, skids=c(10418394,4453485))
}
}
\seealso{
\code{\link{catmaid_POSTJ}}
}

