% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_node_count}
\alias{catmaid_get_node_count}
\title{Get the number of nodes per skeleton}
\usage{
catmaid_get_node_count(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
An integer vector of node counts
}
\description{
Get the number of nodes per skeleton
}
\details{
This is actually a thin wrapper around the
  \code{\link{catmaid_get_review_status}} function, which returns the number
  of reviewed and total nodes for one or more neurons. However this function
  can efficiently return the number of nodes when there are duplicates in the
  input list.
}
\examples{
\dontrun{
catmaid_get_node_count("glomerulus DA2")

# NB handles repeated input efficiently
skids=catmaid_skids("glomerulus DA2 right")
catmaid_get_node_count(rep(skids,20))
}
}
