% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{catmaid_get_review_status}
\alias{catmaid_get_review_status}
\title{Get review status of neurons from CATMAID}
\usage{
catmaid_get_review_status(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
a data.frame consisting of total number of nodes and number of
  reviewed nodes. Note that should any neurons be invalid a warning will be
  generated and a row with NA values will be returned.
}
\description{
Get review status of neurons from CATMAID
}
\examples{
\dontrun{
catmaid_get_review_status(skids=c(10418394,4453485))
}
}
\seealso{
\code{\link{catmaid_fetch}}
}
