% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_connection.R
\name{catmaid_get_server}
\alias{catmaid_get_server}
\title{Get the CATMAID server in use}
\usage{
catmaid_get_server(conn = NULL, ...)
}
\arguments{
\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_login}
}
\description{
shows the URL for a connection object in use (see \code{catmaid_login})
}
\seealso{
\code{\link{catmaid_login}}
}
