% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_treenode_table}
\alias{catmaid_get_treenode_table}
\title{Return tree node table for a given neuron}
\usage{
catmaid_get_treenode_table(skid, pid = 1, conn = NULL, raw = FALSE, ...)
}
\arguments{
\item{skid}{Numeric skeleton id}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
A data.frame with columns \itemize{

  \item id

  \item type

  \item tags

  \item confidence

  \item x

  \item y

  \item z

  \item s

  \item r

  \item user

  \item last_modified

  \item reviewer

  }
}
\description{
Return tree node table for a given neuron
}
\examples{
\dontrun{
# get tree node table for neuron 10418394
tnt=catmaid_get_treenode_table(10418394)
# show all leaf nodes
subset(tnt, type=="L")
# table of node types
table(tnt$type)
}
}
\seealso{
\code{\link{catmaid_get_compact_skeleton}}. \code{\link{read.neuron.catmaid}}
}

