% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_treenodes_detail}
\alias{catmaid_get_treenodes_detail}
\title{Fetch position and other information for arbitrary treenode ids}
\usage{
catmaid_get_treenodes_detail(tnids, pid = 1, conn = NULL, raw = FALSE,
  ...)
}
\arguments{
\item{tnids}{One or more (integer) treenode ids}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE}) 
or to convert the nodes and connectors lists into processed data.frames 
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\value{
\itemize{

  \item treenode_id,

  \item parent_id,

  \item x,

  \item y,

  \item z,

  \item confidence,

  \item radius,

  \item skid,

  \item edition_time,

  \item user_id

  }
}
\description{
Fetch position and other information for arbitrary treenode ids
}
\details{
The key feature of this function is that allows you to fetch
  information for arbitrary tree nodes that do not need to be from the same
  skeleton.
}
\examples{
\donttest{
# details for 3 nodes from two different skeletons
catmaid_get_treenodes_detail(c(9943214L, 25069047L, 12829015L))
}
}
\seealso{
\code{\link{catmaid_get_treenode_table}},
  \code{\link{catmaid_get_connectors}},
  \code{\link{catmaid_get_compact_skeleton}}
}
