% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_stats.R
\name{catmaid_get_user_list}
\alias{catmaid_get_user_list}
\title{Fetch list of catmaid users for current/specified connection/project}
\usage{
catmaid_get_user_list(pid = 1, conn = NULL, ...)
}
\arguments{
\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by 
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection 
object will be generated using the values of the \bold{catmaid.*} package 
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the catmaid_fetch function}
}
\description{
Fetch list of catmaid users for current/specified connection/project
}
\examples{
\dontrun{
catmaid_get_user_list()
}
}
\seealso{
\code{\link{catmaid_get_review_status}, 
  \link{catmaid_get_contributor_stats}}
}

