% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{catmaid_get_volume}
\alias{catmaid_get_volume}
\title{Fetch a mesh volume from CATMAID}
\usage{
catmaid_get_volume(x, rval = c("mesh3d", "catmaidmesh", "raw"), conn = NULL,
  pid = 1, ...)
}
\arguments{
\item{x}{The integer volume id or character volume name}

\item{rval}{The class to return.}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{pid}{project id (default 1)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\value{
When rval \code{mesh3d}, the standard format used by the 
  \code{rgl} and \code{nat} packages is the default. \code{catmaidmesh} is a
  tidy version of the mesh format used by catmaid. \code{raw} will allow any
  return value.
}
\description{
Fetch a mesh volume from CATMAID
}
\details{
Note that if \code{x} is a volume name as a character vector then it
  must be unique (something that CATMAID servers do not insist on).
  
  Note also that this function currently only supports
}
\examples{
\dontrun{
v375=catmaid_get_volume(375)
v13.AL_R=catmaid_get_volume("v13.AL_R")

v13.AL_R.mesh=as.mesh3d(v13.AL_R)
shade3d(v13.AL_R.mesh)

# find surfaces for olfactory glomeruli
vl=catmaid_get_volumelist()
glomids=vl$id[grepl("_glomerulus$", vl$name)]
# fetch them all
gg=lapply(glomids, catmaid_get_volume)
# ... and plot
mapply(shade3d, gg, col=rainbow(length(gg)))

# alternatively wrap them all in an rgl shapelist3d, plotting each object
sl=shapelist3d(gg, col=rainbow(length(gg)), plot=TRUE)

}
}
\seealso{
\code{\link{catmaid_get_volumelist}}, 
  \code{\link{catmaid_add_volume}}, \code{\link{as.catmaidmesh}}, 
  \code{\link[rgl]{mesh3d}}, \code{\link[nat]{hxsurf}},
  \code{\link[rgl]{shapelist3d}}
}
