% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{catmaid_get_volumelist}
\alias{catmaid_get_volumelist}
\title{Fetch a data frame containing details of all volumes on CATMAID server}
\usage{
catmaid_get_volumelist(conn = NULL, pid = 1, ...)
}
\arguments{
\item{conn}{the \code{\link{catmaid_connection}} object}

\item{pid}{project id (default 1)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\value{
A data.frame with columns \itemize{
  
  \item comment
  
  \item name
  
  \item creation_time
  
  \item edition_time
  
  \item project
  
  \item user
  
  \item id
  
  \item editor }
}
\description{
Fetch a data frame containing details of all volumes on CATMAID server
}
\seealso{
\code{\link{catmaid_get_volume}}
}
