% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_islatest}
\alias{catmaid_islatest}
\title{Check if CATMAID skeleton ids are up to date}
\usage{
catmaid_islatest(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One more skids in any form accepted by
\code{\link{catmaid_skids}}. Note that \code{NA}s are accepted and will
give NA output.}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
A logical vector of the same length as the \code{skids} argument.
  Note  duplicates will also still be found in the output.
}
\description{
Check if CATMAID skeleton ids are up to date
}
\details{
NB this will not distinguish between an ID that has been merged away
  and one that never actually existed.
}
\examples{
\donttest{
# for FAFB CATMAID
catmaid_islatest(c(13682, 9254, 15886))
}
}
