% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_latestid}
\alias{catmaid_latestid}
\title{Return the latest skeleton id accounting for edits}
\usage{
catmaid_latestid(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One more skids in any form accepted by
\code{\link{catmaid_skids}}. Note that \code{NA}s are accepted and will
give NA output.}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
A named vector of skeleton ids
}
\description{
Return the latest skeleton id accounting for edits
}
\examples{
\dontrun{
# for FAFB CATMAID
catmaid_latestid(c(13682, 9254, 15886))
}
}
