% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_annotations.R
\name{catmaid_lock_neurons}
\alias{catmaid_lock_neurons}
\alias{catmaid_unlock_neurons}
\title{Lock or unlock a CATMAID neuron reconstruction}
\usage{
catmaid_lock_neurons(skids, pid = 1, conn = NULL, ...)

catmaid_unlock_neurons(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{the skeleton IDs neurons you wish to lock / unlock}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_fetch}
}
\description{
Lock or unlock a CATMAID neuron reconstruction by adding or removing a 'locked' annotation to a set of skeleton IDs (skids). A locked neuron cannot be edited until it is unlocked.
}
