% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catmaid_login}
\alias{catmaid_connection}
\alias{catmaid_login}
\title{Connect/authenticate to a CATMAID server, returning a connection object}
\usage{
catmaid_login(conn = NULL, ..., Force = FALSE)

catmaid_connection(server = getOption("catmaid.server"),
  username = getOption("catmaid.username"),
  password = getOption("catmaid.password"),
  authname = getOption("catmaid.authname"),
  authpassword = getOption("catmaid.authpassword"))
}
\arguments{
\item{conn}{An optional \code{catmaid_connection} connection object.}

\item{...}{Additional arguemnts passed to catmaid_connection}

\item{Force}{Whether to force a new login to the CATMAID server (default
\code{FALSE})}

\item{server}{url of CATMAID server}

\item{username,password}{Your CATMAID username and password.}

\item{authname,authpassword}{The http basicauth username/password that
optionally secures the CATMAID server.}
}
\description{
\code{catmaid_login} allows you to login to a CATMAID server
  specified by a \code{catmaid_connection} object. If such an object is not
  specified, then one is created, by default using \code{options} of the form
  "catmaid.*".

  The connection object returned by \code{catmaid_login} will then be used
  for future requests to the CATMAID server.
}
\details{
After successful login, the \code{catmaid_connection} will contain a
  \code{cookie} field that includes a sessionid that is required for
  subsequent GET/POST operations.
}

