% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_stats.R
\name{catmaid_node_time}
\alias{catmaid_node_time}
\title{Get the UTC creation / edit time for a CATMAID node}
\usage{
catmaid_node_time(id, time = c("creation_time", "edition_time"),
  pid = 1, conn = NULL, ...)
}
\arguments{
\item{id}{a treenode or connector ID}

\item{time}{whether to return the creation_time or edition_time}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see \code{catmaid_login} for details}

\item{...}{methods passed to catmaid_fetch}
}
\description{
Get the UTC creation / edit time for a CATMAID treenode or connector. 
Useful for making 'state' arguments to be passed to other functions that edit data on a CATMAID server.
}
\seealso{
\code{\link{catmaid_get_treenodes_detail}}, \code{\link{catmaid_connector_nodes}}
}
