% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_urls.R
\name{catmaid_parse_url}
\alias{catmaid_parse_url}
\title{Parse a CATMAID URL to extract parameters including xyz location}
\usage{
catmaid_parse_url(x)
}
\arguments{
\item{x}{One or more catmaid urls}
}
\value{
A \code{data.frame} containing parsed parameters. The columns will be
  named according to CATMAID's convention except for the x, y and z location
  (named as such rather than xp, yp, zp as returned by CATMAID).
}
\description{
Parse a CATMAID URL to extract parameters including xyz location
}
\examples{
u=paste0("https://neuropil.janelia.org/tracing/fafb/v12/?pid=1&zp=177100&",
  "yp=166068&xp=373086&tool=tracingtool&sid0=7&s0=1.000000")
catmaid_parse_url(u)

\dontrun{
cpu=catmaid_parse_url(u)
n=read.neuron.catmaid(cpu$active_skeleton_id)
plot3d(n)
spheres3d(xyzmatrix(cpu), r=1000, col='magenta')
}
}
