% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{catmaid_query_by_neuronname}
\alias{catmaid_query_by_neuronname}
\title{Get list of neurons/annotations querying by neuron/annotation name}
\usage{
catmaid_query_by_neuronname(query, pid = 1, maxresults = 500,
  type = c("neuron", "annotation"), raw = FALSE, ...)
}
\arguments{
\item{query}{A query string (NB this can be a regular expression)}

\item{pid}{project id (default 1)}

\item{maxresults}{The maximum number of results to return}

\item{type}{Type of results to return. Defaults to \code{c("neuron",
"annotation")}. Only relevant when \code{raw=FALSE}.}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
a data.frame containing the results with an attribute "annotations"
  containing the annotations as a raw list
}
\description{
Get list of neurons/annotations querying by neuron/annotation name
}
\examples{
\dontrun{
catmaid_query_by_neuronname("ORN")
# using regex functionality
catmaid_query_by_neuronname("ORN (left|right)")
# fancier regex
catmaid_query_by_neuronname("^[0-9a-f &]+ PN (left|right)")
}
}

