% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{catmaid_query_connected}
\alias{catmaid_query_connected}
\title{Find neurons connected to a starting neuron}
\usage{
catmaid_query_connected(skid, minimum_synapses = 1, pid = 1, raw = FALSE,
  ...)
}
\arguments{
\item{skid}{single skeleton id}

\item{minimum_synapses}{Must be at least this number of synapses between
starter neuron and returned partners}

\item{pid}{project id (default 1)}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
A list containing two data.frames (incoming, outgoing), each
  data.frame having one row for each partner neuron and the following columns
  \itemize{

  \item skid skeleton id of the query neuron

  \item partner the skeleton id of the partner neuron

  \item syn.count the number of synapses made with partner

  \item num_nodes number of nodes in the partner neuron

  }
}
\description{
Find neurons connected to a starting neuron
}
\examples{
\dontrun{
orn13a=catmaid_query_by_name("13a ORN left")$skid
catmaid_query_connected(orn13a)
}
}
\seealso{
\code{\link{catmaid_get_review_status}} to get information about the
review status of partners (as shown in the equivalent CATMAID report).
}

