% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_connection.R
\name{catmaid_record_markdown}
\alias{catmaid_record_markdown}
\title{Record and use a web cache when running an R script / Rmarkdown file}
\usage{
catmaid_record_markdown(path, ...)
}
\arguments{
\item{path}{Optional path to location to cache http requests}

\item{...}{Additional arguments passed to
\code{\link[httptest]{start_vignette}}}
}
\value{
The path to the recording location (invisibly)
}
\description{
Record and use a web cache when running an R script / Rmarkdown file
}
\details{
\code{catmaid_record_markdown} will set up a web cache using the
  \code{\link[httptest]{httptest}} package. This will record all web requests
  made via the httr package for the remainder of the session, including all
  calls made via the \bold{catmaid} package to CATMAID servers. If you are
  knitting a markdown file the session lasts for that file only. If you are
  running interactively, then the session will last until you close R or call
  \code{\link[httptest]{end_vignette}()}. See
  \url{https://cran.r-project.org/web/packages/httptest/vignettes/vignettes.html}
   for further details of  the httptest package.

  By default when you are using an Rstudio project, the http cache will be a
  folder called \code{catmaid_recording} placed next to your \code{Rproj}
  file. If you are knitting an Rmarkdown file that is not associated with an
  Rstudio project, the recording path for \code{myfile.rmd} will be
  \code{myfile.catmaid_recording}.
}
\examples{

\dontrun{
# You can use this when you have an Rstudio project open
library(httptest)
catmaid_record_markdown()

# You can place the following towards the start of your Rmarkdown file
```{r, include=FALSE}
library(httptest)
catmaid::catmaid_record_markdown()
```
# If you intend to run your Rmarkdown interactively, it is recommended to add
# the following at the end of your Rmarkdown.
# see https://cran.r-project.org/web/packages/httptest/vignettes/vignettes.html
# for details
```{r, include=FALSE}
end_vignette()
```
}
}
\seealso{
\code{\link[httptest]{httptest}},
  \code{\link[httptest]{start_vignette}},
  \code{\link[httptest]{end_vignette}}
}
