% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{catmaid_rename_neuron}
\alias{catmaid_rename_neuron}
\title{Rename (presently) a single neuron or other entity (e.g. annotation)}
\usage{
catmaid_rename_neuron(skids = NULL, entityids = NULL, names, pid = 1,
  conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{entityids}{Generic entity ids}

\item{names}{New names}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
\code{logical} value indicating success or failure with attributes 
  indicating the previous name.
}
\description{
Rename (presently) a single neuron or other entity (e.g. annotation)
}
\details{
Each neuron consists of both a generic entity and an associated 
  skeleton (\code{skid}). If you supply the \code{skid} the function will 
  internally look up the corresponding neuron entity id. Note that you 
  actually wished to change the name of the skeleton you could pass the skid 
  into the \code{entityids} argument.
}
\examples{
\dontrun{
# Warning this will change data!
n=get_neuron_name(27296)
catmaid_rename_neuron(skids=27296, names=n)
}
}
