% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_search.R
\name{catmaid_skeletons_in_bbx}
\alias{catmaid_skeletons_in_bbx}
\title{Search for CATMAID skeletons within a volume}
\usage{
catmaid_skeletons_in_bbx(bbx, min_nodes = 2, pid = 1, conn = NULL,
  ...)
}
\arguments{
\item{bbx}{the bounding box (a matrix of 2 rows and 3 columns) describing a search volume}

\item{min_nodes}{the minimum number of nodes a neuron in the search area must have (includes nodes outside search area)}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid_login for details}

\item{...}{methods passed to catmaid_fetch}
}
\description{
Programmatically search for skeleton IDs pertaining to neurons within a search volume defined by a bounding box.
}
