% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catmaid_skids.R
\name{catmaid_skids}
\alias{catmaid_skids}
\title{Find skeleton ids (skids) for various inputs including textual queries}
\usage{
catmaid_skids(x, several.ok = TRUE)
}
\arguments{
\item{x}{one or more skids or a query expression (see details)}

\item{several.ok}{Logical}
}
\value{
\code{integer} vector of skids (of length 0 on failure).
}
\description{
Find skeleton ids (skids) for various inputs including textual queries
}
\details{
If the inputs are numeric or have length > 1 they are assumed
  already to be skids and are simply converted to integers. If the the input
  is a string starting with "name:" or "annotation:" they are used for a
  query by \code{\link{catmaid_query_by_name}} or
  \code{\link{catmaid_query_by_annotation}}, respectively.
}

