% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_nat.R
\name{connectors}
\alias{connectors}
\alias{connectors.catmaidneuron}
\alias{connectors.neuronlist}
\title{Get data.frame of connector (synapse) information from a neuron or \code{neuronlist}}
\usage{
connectors(x, ...)

\method{connectors}{catmaidneuron}(x, ...)

\method{connectors}{neuronlist}(x, subset = NULL, ...)
}
\arguments{
\item{x}{Neuron or neuronlist}

\item{...}{Additional arguments passed to methods (and to \code{nlapply} in 
the case of connectors)}

\item{subset, }{optional subset of neurons to keep (see
\code{\link[nat]{nlapply}} for details)}
}
\value{
A data.frame with columns \itemize{
  
  \item treenode_id (NB this is the treenode id for the \emph{current} skeleton)
  
  \item connector_skid
  
  \item prepost integer indicating whether connection is pre-(\code{0}) or 
  post(\code{1})-synaptic with respect to the current neuron. In other words 
  this field will be \code{0} (pre) for the output synapses of this neuron.
  
  \item x Spatial Location
  
  \item y
  
  \item z
  
  \item skid For \code{connectors.neuronlist}, the skid of the skeleton from
  which connector information was retrieved.
  
  }
}
\description{
In contrast to \code{\link{catmaid_get_connector_table}} this assumes that
you have already read the neurons into an R structure of class 
\code{\link[nat]{neuron}} or \code{\link[nat]{neuronlist}}.
}
\details{
Note that the id column identifying each neuron will be called
  \code{bodyid} or \code{skid} if such a field exists in the metadata
  attached to the neuron or \code{id} otherwise. If the column is called
  skid, the identifier will be converted to an integer otherwise it will be a
  character vector. These adjustments will avoid large 64 bit int ids from
  neuprint being zapped.
}
\examples{
\dontrun{
ornsl=read.neurons.catmaid("name:ORN left", OmitFailures = T, .progress='text')
conndf=connectors(ornsl)
summary(connectors(ornsl))

# plot points in 3d
library(nat)
nopen3d()
points3d(xyzmatrix(conndf), col=c(pre='red', post='cyan')[conndf$prepost+1])
}
}
\seealso{
\code{\link[nat]{nlapply}}

Other connectors: 
\code{\link{catmaid_adjacency_matrix}()},
\code{\link{catmaid_get_connector_table}()},
\code{\link{catmaid_get_connectors_between}()},
\code{\link{catmaid_get_connectors}()},
\code{\link{catmaid_query_connected}()}
}
\concept{connectors}
