% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_nat.R
\name{copy_tags_connectors}
\alias{copy_tags_connectors}
\title{Copy the tag / connector info from a catmaid neuron to a new neuron}
\usage{
copy_tags_connectors(new, old, update_node_ids = TRUE)
}
\arguments{
\item{new, old}{information will be copied from old -> new}

\item{update_node_ids}{whether to update the node ids (default \code{TRUE})}
}
\value{
A new neuron containing tag and connector information from the old 
  neuron and having the same class as the old neuron.
}
\description{
This function is intended primarily for developer use as a way 
  to copy over connector/tag information typically contained in CATMAID
  neurons but lost when nat functions are applied to transform a neuron.
}
\details{
Both connectors and tag information contain node ids. These may need
  to be mapped onto new node ids if these are not consistent between old and 
  new neurons.
}
