% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_nat.R
\name{nsoma}
\alias{nsoma}
\alias{nsoma.neuronlist}
\alias{nsoma.neuron}
\alias{nsoma.default}
\title{Return the number of explicitly tagged somata in a (CATMAID) neuron}
\usage{
nsoma(x, ...)

\method{nsoma}{neuronlist}(x, ...)

\method{nsoma}{neuron}(x, ...)

\method{nsoma}{default}(x, soma_label = "soma", ...)
}
\arguments{
\item{x}{Objects to count somata e.g. one or more neurons or a specifier
passed to \code{\link{catmaid_skids}}}

\item{...}{Additional arguments, eventually passed by \code{nsoma.default} to
\code{\link{catmaid_skids}}, otherwise ignored.}

\item{soma_label}{Character vector of one or more label names that identify
somata.}
}
\value{
A named integer vector corresponding to the number of neurons
  specified by \code{x}.
}
\description{
Return the number of explicitly tagged somata in a (CATMAID) neuron
}
\details{
These functions can cope with loaded neuron objects or CATMAID skid
  specifications (see \code{\link{catmaid_skids}}). Note that this function
  will return 0 for any neuron that does not contain a \code{tags$soma}
  entry, including regular \code{\link{neuron}} objects .
}
\examples{
nsoma(Cell07PNs)
data("AV4b1")
nsoma(AV4b1)
\donttest{
nsoma("ORN PNs")
}
\dontrun{
nsoma(3486381)
}
}
\seealso{
\code{\link{soma}}
}
