% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_tags.R
\name{prune_by_tag}
\alias{prune_by_tag}
\title{Prune neuron by splitting it at CATMAID tags}
\usage{
prune_by_tag(x, tag, remove.upstream = TRUE, ...)
}
\arguments{
\item{x}{a \code{neuron} or \code{neuronlist} object}

\item{tag}{a tag that has been assigned in CATMAID}

\item{remove.upstream}{Logical when \code{TRUE} points downstream of the tag(s) are removed, if true, points upstream are removed}

\item{...}{Additional arguments, passed to \code{nlapply} or eventually to \code{prune_vertices}}
}
\description{
Prune neuron by splitting it at CATMAID tags
}
\details{
Split a neuron based on tags assigned in CATMAID. Remove points either downstream (from the root, must be soma to work properly) of the tagged points(s) or upstream.
}
