% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_selection.R
\name{read_catmaid_selection}
\alias{read_catmaid_selection}
\title{Read a CATMAID selection file}
\usage{
read_catmaid_selection(f, readNeurons = FALSE, getNames = TRUE, ...)
}
\arguments{
\item{f}{Path to the \code{.json} file saved by CATMAID}

\item{readNeurons}{Whether or not to read the neurons listed in the selection
file}

\item{getNames}{Whether to fetch the names of the neurons (default 
\code{TRUE}, ignored if neurons are being fetched)}

\item{...}{Connection parameters passed to \code{read.neurons.catmaid}}
}
\value{
A \code{data.frame} containing the selection information or (if 
  \code{readNeurons=TRUE}) a \code{\link[nat]{neuronlist}} containing this 
  information in the attached metadata. The columns will be \itemize{
  
  \item{skid} The numeric skeleton id
  
  \item{color} The colour of the selected neuron in catmaid
  
  \item{opacity} The opacity (alpha value) of the neuron name
  
  \item{name} Optionally the name of the neuron
  
  }
}
\description{
Read a CATMAID selection file
}
\examples{
f=system.file('catmaid-skeletons.json', package = 'catmaid')
read_catmaid_selection(f, getNames=FALSE)
\dontrun{
# read in the neurons for the selection
x=read_catmaid_selection(f, readNeurons=TRUE)
# plot using the CATMAID colour specifications
plot3d(x, col=color)
}
}
\seealso{
\code{\link{write_catmaid_selection}}
}
