% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_tags.R
\name{soma}
\alias{soma}
\alias{somaid}
\alias{somapos}
\title{Functions for retrieving soma data}
\usage{
soma(x, ...)

somaid(x, ...)

somapos(x, ...)
}
\arguments{
\item{x}{a neuron or neuronlist object}

\item{...}{additional arguments passed to methods, i.e. \code{nlapply}}
}
\value{
Either the soma's 3D coordinates (soma) its id number (somaid) or its position in the neuron's skeleton (somapos)
}
\description{
Functions for retrieving data about somata from skeletons, that has been assigned in CATMAID using the tag system
}
\details{
CATMAID access required. Data collected and described in cited publication.
}
\references{
Ohyama T, Schneider-Mizell CM, Fetter RD, Aleman JV, Franconville R, Rivera-Alba M, Mensh BD, Branson KM, Simpson JH, Truman JW, et al. (2015) A multilevel multimodal circuit enhances action selection in Drosophila. Nature.
}
