% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{update_neurons}
\alias{update_neurons}
\alias{update_names}
\title{Update a local neuronlist with new CATMAID data}
\usage{
update_neurons(someneuronlist, skids, ...)

update_names(someneuronlist, skids = names(someneuronlist), ...)
}
\arguments{
\item{someneuronlist}{a neuronlist object}

\item{skids}{skeleton IDs}

\item{...}{additional arguments passed to \code{read.neurons.catmaid}}
}
\description{
Use to update a large neuronlist quickly by pulling just certain neurons from CATMAID,
using \code{read.neurons.catmaid}.
}
