% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_selection.R
\name{write_catmaid_selection}
\alias{write_catmaid_selection}
\title{Make a CATMAID selection file based on neuronlist or skids}
\usage{
write_catmaid_selection(x, f, color = NULL, opacity = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{neuronlist}}, data.frame or vector of skids}

\item{f}{Path to output file (usually ends in .json)}

\item{color}{Optional vector of colours in any format understood by
\code{\link{col2rgb}}}

\item{opacity}{Optional vector of opacities (alpha values) in range 0-1}

\item{...}{Additional arguments passed to \code{\link{catmaid_skids}}}
}
\description{
Make a CATMAID selection file based on neuronlist or skids
}
\examples{
\dontrun{
# Write selection file for neurons defined by annotation or name
write_catmaid_selection("PN", "pns.json")
write_catmaid_selection("name:AV6", "name:av6.json")

pns=read.neurons.catmaid("PN")
# extract the glomerulus from the name
pns[, 'glomerulus'] = stringr::str_match(pns[, 'name'],
                                         ".*glomerulus ([A-z0-9]+) .*")[, 2]
pns[, 'glomerulus']=addNA(factor(pns[, 'glomerulus']))
# get the same colours that nat would normally use in a plot
plotres=plot3d(pns, col=glomerulus)
cols=attr(plotres,'df')$col
# write out selection file with those colours
write_catmaid_selection(pns, f='pns-by-glom.json', color=cols)
}
}
\seealso{
\code{\link{read_catmaid_selection}}
}
