% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{catmaid-package}
\alias{catmaid}
\alias{catmaid-package}
\title{Programmatic access to the CATMAID API}
\description{
To use, first call \code{\link{catmaid_login}} to authenticate to a catmaid
server. You can then use functions such as
\code{\link{catmaid_get_neuronnames}} or
\code{\link{catmaid_get_compact_skeleton}} to query the server. You can also
roll your own server queries using the low-level \code{\link{catmaid_fetch}}
command.
}
\section{Connections}{
 By default all query functions will use the
  server/authentication details specified by the last successful call to
  \code{\link{catmaid_login}} unless a \code{\link{catmaid_connection}}
  object is provided as an argument.
}

\section{Package Options}{
 See \code{\link{catmaid_login}} for details of the
  options that can be set to specify default login values. It may make sense
  to set these in your \code{.Rprofile} if you regularly use a specific
  CATMAID server. Beware however that your R workspace will then include a
  plain text version of your passwords.
}

\section{Unit tests}{
 Unit tests can be run as per the examples section below.
  In order to do live tests you will need to set environment variables with
  your catmaid login credentials. This is detailed in file
  \code{testthat/test-connections.R}.
}
\examples{
\dontrun{
## test package
library(catmaid)
library(testthat)
test_package("catmaid")
}
}
\references{
Based in large part on code visible at
  \url{https://github.com/schlegelp/CATMAID-to-Blender/blob/master/CATMAIDImport.py}
   See \url{http://catmaid.org} for further details about CATMAID and
  \url{https://github.com/acardona/CATMAID/blob/master/django/applications/catmaid/urls.py}
   for a list of URLs that the web API accepts.
}
\keyword{package}

