% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid-package.R, R/catmaid.R
\docType{package}
\name{catmaid-package}
\alias{catmaid}
\alias{catmaid-package}
\title{catmaid: R access to the API for the CATMAID web image annotation
    tool}
\description{
This package provide access to the API exposed
    by the CATMAID (C ollaborative A nnotation T oolkit for M assive A
    mounts of I mage D ata) web tool. This includes a bridge to the
    neuroanatomy toolbox (R package nat), thereby allowing analysis and
    visualisation of neuronal skeletons.

To use, first call \code{\link{catmaid_login}} to authenticate to a catmaid
server. You can then use functions such as
\code{\link{catmaid_get_neuronnames}} or
\code{\link{catmaid_get_compact_skeleton}} to query the server. You can also
roll your own server queries using the low-level \code{\link{catmaid_fetch}}
command.
}
\section{Connections}{
 By default all query functions will use the
  server/authentication details specified by the last successful call to
  \code{\link{catmaid_login}} unless a \code{\link{catmaid_connection}}
  object is provided as an argument.
}

\section{Default Login}{
 See \code{\link{catmaid_login}} for details of the
  environment variables that can be set to specify default login values. It
  makes sense to set these in your \code{\link{.Renviron}} file if you
  regularly use a specific CATMAID server. Note that you can have different
  \code{.Renviron} files in the home folder of different Rstudio projects if
  you regularly use more than one CATMAID server. Note that the use of 
  environment variables is now preferred over setting R options.
}

\section{Unit tests}{
 Unit tests can be run as per the examples section below.
  In order to do live tests you will need to set environment variables with
  your catmaid login credentials using the function
  \code{\link{catmaid_connection_setenv}}.
}

\examples{
\dontrun{
## test package
library(catmaid)
library(testthat)
test_package("catmaid")

## same but use appropriate login info
conn=catmaid_login(server="https://myserver.com", user="calvin", password='hobbes')
## alternatively, login using an API token
# see http://catmaid.github.io/dev/api.html#api-token for how to get one
conn=catmaid_login(server="https://myserver.com",
                   token = "9944b09199c62bcf9418ad846dd0e4bbdfc6ee4b")
catmaid_connection_setenv(conn)
test_package("catmaid")
}
}
\references{
The first version of this package was based in large part on
  python code at
  \url{https://github.com/schlegelp/CATMAID-to-Blender/blob/master/CATMAIDImport.py}.
   See \url{http://catmaid.org} for further details about CATMAID and
  \url{https://github.com/acardona/CATMAID/blob/master/django/applications/catmaid/urls.py}
   for a list of URLs that the web API accepts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/rcatmaid}
  \item Report bugs at \url{https://github.com/jefferis/rcatmaid/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

Other contributors:
\itemize{
  \item Sridhar Jagannathan \email{j.sridharrajan@gmail.com} (0000-0002-2078-1145) [contributor]
  \item Alex Bates \email{abates@mrc-lmb.cam.ac.uk} (0000-0002-1195-0445) [contributor]
}

}
\keyword{internal}
\keyword{package}
