% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_annotations.R
\name{catmaid_entities_from_models}
\alias{catmaid_entities_from_models}
\title{Return the entity ids for one or more model ids}
\usage{
catmaid_entities_from_models(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
An integer vector of \bold{entity ids} each named by the 
  corresponding \bold{model id} (usually a skeleton id).
}
\description{
Return the entity ids for one or more model ids
}
\details{
This will normally be used to turn skeleton ids into neuron ids 
  which are used e.g. for annotation purposes. This is probably not something
  that many end users will need but is required e.g. by 
  \code{catmaid_remove_annotations_for_skeletons}.
}
\seealso{
Used by \code{\link{catmaid_remove_annotations_for_skeletons}}
}
