% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catmaid_GET}
\alias{catmaid_GET}
\alias{catmaid_GETJ}
\title{Send http GET request to a CATMAID server}
\usage{
catmaid_GET(path, conn = NULL, ...)

catmaid_GETJ(path, conn = NULL, ...)
}
\arguments{
\item{path}{The path on the CATMAID server relative to the CATMAID root}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{httr::GET} function}
}
\value{
For \code{catmaid_GET} an object of class \code{response} or, for for
  \code{catmaid_GETJ}, the raw R object generated by calling
  \code{jsonlite::\link[jsonlite]{fromJSON}} on the body of the response.
}
\description{
\code{catmaid_GET} returns the raw response, whereas
  \code{catmaid_GETJ} parses JSON content into an R object.
}
\examples{
\dontrun{
conn=catmaid_login()
## fetch a demo skeleton
# raw response
skel.response=catmaid_GET("1/10418394/0/0/compact-skeleton", conn)
# list object
skel=catmaid_GETJ("1/10418394/0/0/compact-skeleton", conn)
}
}
\seealso{
\code{\link{catmaid_login}}, \code{\link[httr]{GET}}
}

