% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_annotations.R
\name{catmaid_get_annotations_for_skeletons}
\alias{catmaid_get_annotations_for_skeletons}
\alias{catmaid_set_annotations_for_skeletons}
\alias{catmaid_remove_annotations_for_skeletons}
\title{Get or set annotations for neurons from CATMAID}
\usage{
catmaid_get_annotations_for_skeletons(skids, pid = 1, conn = NULL, ...)

catmaid_set_annotations_for_skeletons(skids, annotations, pid = 1,
  conn = NULL, ...)

catmaid_remove_annotations_for_skeletons(skids, annotations,
  force = FALSE, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}

\item{annotations}{Character vector of one or more named annotations to add 
to the specified neurons.}

\item{force}{Whether to force the catmaid server to remove multiple 
annotations (default \code{FALSE}) to provide some protection against 
accidents.}
}
\value{
For \code{catmaid_get_annotations_for_skeletons} a data.frame 
  containing the following columns \itemize{
  
  \item skid The skeleton id
  
  \item annotation The annotation string
  
  \item id The annotation id
  
  \item uid The user id for the annotation
  
  }
  
  For \code{catmaid_set_annotations_for_skeletons} a list containing 
  information about the annotations that have just been added.
For \code{catmaid_remove_annotations_for_skeletons} a list containing
information about the annotations that have just been removed.
}
\description{
Pretty obviously: \code{catmaid_get_annotations_for_skeletons}
  gets annotations from one or more neurons.

Pretty obviously: \code{catmaid_set_annotations_for_skeletons} 
  sets annotations for one or more neurons. Although adding annotations is 
  non-destructive, \bold{please use carefully} since this will be making
  changes on the server!
}
\details{
Note that annotations will be created on the catmaid server if they 
  do not already exist (and will not be re-created if they do already exist.)
}
\examples{
\dontrun{
catmaid_get_annotations_for_skeletons(skids=c(10418394,4453485))
catmaid_get_annotations_for_skeletons("name:ORN (left|right)")
catmaid_get_annotations_for_skeletons("annotation:ORN PNs$")
}
\dontrun{
catmaid_set_annotations_for_skeletons(skids=c(10418394,4453485), 'myselection')
catmaid_set_annotations_for_skeletons(skids="annotation:ORN PNs$", 'my pns')
}
\dontrun{
nn=c(10418394,4453485)
catmaid_set_annotations_for_skeletons(skids=nn, 'mytest')
catmaid_remove_annotations_for_skeletons(nn, 'mytest')
}
}
\seealso{
\code{\link{catmaid_query_by_annotation}},
  \code{\link{catmaid_get_meta_annotations}} \code{\link{catmaid_fetch}},
  \code{\link{catmaid_skids}}
}
