% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catmaid_get_compact_skeleton}
\alias{catmaid_get_compact_skeleton}
\title{return the raw data for a CATMAID neuronal skeleton}
\usage{
catmaid_get_compact_skeleton(pid, skid, conn = NULL, connectors = TRUE,
  tags = TRUE, raw = FALSE, ...)
}
\arguments{
\item{pid,skid}{project and skeleton ids}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{connectors}{Whether to fetch connector information}

\item{tags}{Whether to fetch tag information}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\description{
return the raw data for a CATMAID neuronal skeleton
}
\examples{
\dontrun{
## using an existing connection object
conn=catmaid_login()
skel=catmaid_get_compact_skeleton(1, 10418394, conn)
# no connector (i.e. synapse) information
skel=catmaid_get_compact_skeleton(1, 10418394, conn, connectors = FALSE)

## using throwaway connection object
skel=catmaid_get_compact_skeleton(1, 10418394)
}
}
\seealso{
\code{link{catmaid_fetch}}
}

