% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_compact_skeleton}
\alias{catmaid_get_compact_skeleton}
\title{Return the raw data for a CATMAID neuronal skeleton}
\usage{
catmaid_get_compact_skeleton(skid, pid = 1L, conn = NULL,
  connectors = TRUE, tags = TRUE, raw = FALSE, ...)
}
\arguments{
\item{skid}{single skeleton id}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{connectors}{Whether to fetch connector information}

\item{tags}{Whether to fetch tag information}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE}) 
or to convert the nodes and connectors lists into processed data.frames 
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\description{
Return the raw data for a CATMAID neuronal skeleton
}
\details{
Note that by default this fetches both skeleton and connector 
  (synapse) information.
}
\examples{
\dontrun{
## ensure that you have done something like
# conn=catmaid_login()
# at least once this session to connect to the server
skel=catmaid_get_compact_skeleton(10418394)
# no connector (i.e. synapse) information
skel=catmaid_get_compact_skeleton(10418394, connectors = FALSE)

}
}
\seealso{
\code{\link{read.neuron.catmaid}} to read as neuroanatomy toolbox
  neuron that can be plotted directly. \code{\link{catmaid_fetch}}.
}
