% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_connectors}
\alias{catmaid_get_connectors}
\title{Return skeleton ids for pre/postsynaptic partners of a set of connector_ids}
\usage{
catmaid_get_connectors(connector_ids, pid = 1, conn = NULL, raw = FALSE,
  ...)
}
\arguments{
\item{connector_ids}{Numeric ids for each connector (synapse).}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE}) 
or to convert the nodes and connectors lists into processed data.frames 
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\value{
A data.frame with columns \itemize{
  
  \item connector_id
  
  \item pre
  
  \item post
  
  }
}
\description{
Return skeleton ids for pre/postsynaptic partners of a set of connector_ids
}
\details{
Note that this returns pairwise connections between neurons. A 
  single synapse (i.e. connector) may have multiple connections; most 
  commonly a single presynaptic cell connects to multiple post-synaptic
  cells but many variations are possible
}
\seealso{
Other connectors: \code{\link{catmaid_get_connector_info}},
  \code{\link{catmaid_get_connector_table}},
  \code{\link{catmaid_query_connected}},
  \code{\link{connectors}}
}

