% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_stats.R
\name{catmaid_get_contributor_stats}
\alias{catmaid_get_contributor_stats}
\title{Get contributor statistics for neurons from CATMAID}
\usage{
catmaid_get_contributor_stats(skids, pid = 1, conn = NULL, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
a list containing different statistics including construction and 
  review times (aggregated across all the specified input neurons). There 
  will also be 3 data.frames containing statistics for number of nodes and 
  pre/post-synaptic connectors broken down per user.
  
  \itemize{
  
  \item pre_contributors number of pre-synaptic connectors contributed per
  user.
  
  \item node_contributors number of skeleton nodes contributed per user.
  
  \item post_contributors number of post-synaptic connectors contributed per
  user.
  
  }
}
\description{
Get contributor statistics for neurons from CATMAID
}
\examples{
\dontrun{
cs=catmaid_get_contributor_stats(skids=c(10418394,4453485))
# fetch user list
ul=catmaid_get_user_list()
# merge with list of node contributors and sort in descending order of
# contributions
library(dplyr)
left_join(cs$node_contributors, ul) \%>\%
  select(id,n, full_name) \%>\%
  arrange(desc(n))
}
}
\seealso{
\code{\link{catmaid_get_review_status}}
}
