% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_labels.R
\name{catmaid_get_label_stats}
\alias{catmaid_get_label_stats}
\alias{catmaid_get_all_labels}
\title{Get statistics for all labels in a project from a CATMAID server}
\usage{
catmaid_get_label_stats(pid = 1, conn = NULL, ...)

catmaid_get_all_labels(pid = 1, conn = NULL, ...)
}
\arguments{
\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\description{
\code{catmaid_get_label_stats} returns a data.frame with one row
  for every node-tag pair.

\code{catmaid_get_all_labels} returns a character vector of labels
}
\examples{
\donttest{
label_stats=catmaid_get_label_stats()
library(dplyr)
# select soma labels
soma_labels=label_stats \%>\%
  filter(labelName=='soma') \%>\%
  group_by(skeletonID)

# select skeleton ids for neurons with multiple cell bodies
multiple_soma=soma_labels \%>\%
  count(skeletonID) \%>\%
  filter(n>1) \%>\%
  arrange(desc(n))

multiple_soma_info = soma_labels \%>\%
  filter(skeletonID\%in\% multiple_soma$skeletonID)
}
\donttest{
head(catmaid_get_all_labels())
}
}
\seealso{
\code{\link{catmaid_get_review_status}}

Other labels: \code{\link{catmaid_get_labels}}
}
\concept{labels}
