% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{catmaid_get_neuronnames}
\alias{catmaid_get_neuronnames}
\title{Get names of neurons from CATMAID}
\usage{
catmaid_get_neuronnames(skids, pid = 1, ...)
}
\arguments{
\item{skids}{One or more skeleton ids}

\item{pid}{Project id (default 1)}

\item{...}{Additional arguments passed to the catmaid_fetch function}
}
\value{
a named character vector of neuron names. Missing values will be
  represented by a \code{NA_character} value.
}
\description{
Get names of neurons from CATMAID
}
\examples{
\dontrun{
catmaid_get_neuronnames(skids=c(10418394,4453485))
}
}
\seealso{
\code{\link{catmaid_fetch}}
}

