% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skeleton.R
\name{catmaid_get_treenode_table}
\alias{catmaid_get_treenode_table}
\title{Return tree node table for a given neuron}
\usage{
catmaid_get_treenode_table(skid, pid = 1, conn = NULL, raw = FALSE,
  ...)
}
\arguments{
\item{skid}{Numeric skeleton id}

\item{pid}{project id (default 1)}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
A data.frame with columns \itemize{
  
  \item id
  
  \item parent_id
  
  \item confidence
  
  \item x
  
  \item y
  
  \item z
  
  \item r
  
  \item user_id
  
  \item last_modified
  
  \item reviewer_id (character vector with comma separated reviewer ids)
  
  }
  
  In addition two data.frames will be included as attributes: \code{reviews},
  \code{tags}.
}
\description{
Return tree node table for a given neuron
}
\examples{
\dontrun{
# get tree node table for neuron 10418394
tnt=catmaid_get_treenode_table(10418394)

# look at tags data
str(attr(tnt, 'tags'))
# merge with main node table to get xyz position
tags=merge(attr(tnt, 'tags'), tnt, by='id')
# label up a 3d neuron plot
n=read.neuron.catmaid(10418394)
plot3d(n, WithNodes=F)
text3d(xyzmatrix(tags), texts = tags$tag, cex=.7)
}
}
\seealso{
\code{\link{catmaid_get_compact_skeleton}}, 
  \code{\link{read.neuron.catmaid}} and \code{\link{catmaid_get_user_list}} 
  to translate user ids into names.
}
