% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_stats.R
\name{catmaid_get_user_list}
\alias{catmaid_get_user_list}
\alias{catmaid_userids}
\title{Fetch or translate catmaid users for current/specified connection/project}
\usage{
catmaid_get_user_list(pid = 1, conn = NULL, ...)

catmaid_userids(x, pid = 1, conn = NULL, ...)
}
\arguments{
\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}

\item{x}{A character vector of login names. If \code{x} contains valid
integer (or numeric) ids then they will be returned as is.}
}
\value{
For \code{catmaid_get_user_list} a \code{data.frame} with

For \code{catmaid_userids} an integer vector of numeric ids
}
\description{
\code{catmaid_userids} convert login names to numeric ids
}
\examples{
\donttest{
head(catmaid_get_user_list())
}
\donttest{
catmaid_userids(1)
catmaid_userids('greg')
}
}
\seealso{
\code{\link{catmaid_get_review_status}, 
  \link{catmaid_get_contributor_stats}}
}
