% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{catmaid_get_volume}
\alias{catmaid_get_volume}
\title{Fetch a mesh volume from CATMAID}
\usage{
catmaid_get_volume(x, rval = c("mesh3d", "catmaidmesh", "raw"),
  invertFaces = FALSE, conn = NULL, pid = 1, ...)
}
\arguments{
\item{x}{The integer volume id or character volume name}

\item{rval}{The class to return.}

\item{invertFaces}{Whether to invert the faces (swapping inside and outside, 
see details).}

\item{conn}{the \code{\link{catmaid_connection}} object}

\item{pid}{project id (default 1)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\value{
When rval \code{mesh3d}, the standard format used by the \code{rgl} 
  and \code{nat} packages is the default. \code{catmaidmesh} is a tidy 
  version of the mesh format used by catmaid. \code{raw} will allow any 
  return value.
}
\description{
Fetch a mesh volume from CATMAID
}
\details{
Note that if \code{x} is a volume name as a character vector then it
  must be unique (something that CATMAID servers do not insist on).
  
  Note also that this function is targeted at mesh volumes, but CATMAID also 
  support box volumes - these can still be returned and will have a valid set
  of vertices but an invalid set of indices.
  
  I have noticed that some CATMAID meshes are returned with the normals
  facing into the mesh (the opposite of what one would expect). This causes
  \code{\link[nat]{pointsinside}} to do the opposite of what one would
  expect. Use \code{invertFaces=TRUE} to fix this.
}
\examples{
## NB all these examples refer to the FAFB adult Drosophila brain
## CATMAID instance
\donttest{
v14.neuropil=catmaid_get_volume('v14.neuropil')
# specifying by name is easier / less fragile than numeric ids
}

\dontrun{
AL_R=catmaid_get_volume("AL_R")
shade3d(AL_R, col='red', alpha=.3)

if(require("Morpho")) {
  plotNormals(facenormals(v13.AL_R), long=5e3)
}

# find surfaces for olfactory glomeruli
vl=catmaid_get_volumelist()
x
glomids=vl$id[grepl("v14.[DV][MLAPC]{0,1}[0-9]{0,1}[dvaplm]{0,1}$", vl$name)]
# fetch them all
gg=lapply(glomids, catmaid_get_volume)
# ... and plot
mapply(shade3d, gg, col=rainbow(length(gg)))

# alternatively wrap them all in an rgl shapelist3d, plotting each object
sl=shapelist3d(gg, col=rainbow(length(gg)), plot=TRUE)

}
}
\seealso{
\code{\link{catmaid_get_volumelist}}, 
  \code{\link{catmaid_add_volume}}, \code{\link{as.catmaidmesh}}, 
  \code{\link[rgl]{mesh3d}}, \code{\link[nat]{hxsurf}}, 
  \code{\link[rgl]{shapelist3d}}
}
