% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{catmaid_get_volumelist}
\alias{catmaid_get_volumelist}
\title{Fetch a data frame containing details of all volumes on CATMAID server}
\usage{
catmaid_get_volumelist(conn = NULL, pid = 1, ...)
}
\arguments{
\item{conn}{the \code{\link{catmaid_connection}} object}

\item{pid}{project id (default 1)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}} 
function.}
}
\value{
Since \code{CATMAID v2018.07.19-457-ga09910b} a \code{data.frame}
  with columns \itemize{

  \item id

  \item name

  \item comment

  \item user_id

  \item editor_id

  \item project_id

  \item creation_time

  \item edition_time

  \item annotations

  }
}
\description{
Fetch a data frame containing details of all volumes on CATMAID server
}
\examples{
\donttest{
vl <- catmaid_get_volumelist()
head(vl)
# find meshes whose name matches neuropil
subset(vl, grepl('neuropil', name))
}
}
\seealso{
\code{\link{catmaid_get_volume}}
}
