% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catmaid_login}
\alias{catmaid_connection}
\alias{catmaid_login}
\title{Connect/authenticate to a CATMAID server, returning a connection object}
\usage{
catmaid_login(conn = NULL, ..., Force = FALSE)

catmaid_connection(server = getOption("catmaid.server"),
  username = getOption("catmaid.username"),
  password = getOption("catmaid.password"),
  authname = getOption("catmaid.authname"),
  authpassword = getOption("catmaid.authpassword"),
  authtype = getOption("catmaid.authtype", default = "basic"))
}
\arguments{
\item{conn}{An optional \code{catmaid_connection} connection object.}

\item{...}{Additional arguemnts passed to catmaid_connection}

\item{Force}{Whether to force a new login to the CATMAID server (default
  \code{FALSE})}

\item{server}{url of CATMAID server}

\item{username,password}{Your CATMAID username and password.}

\item{authname,authpassword}{The http username/password that optionally
secures the CATMAID server.}

\item{authtype}{The http authentication scheme, see
\code{\link[httr]{authenticate}} for details.}
}
\value{
a \code{catmaid_connection} object that can be used to make
  authenticated http requests to a CATMAID server, specifically by making
  use of the \code{$config} field.
}
\description{
\code{catmaid_login} allows you to login to a CATMAID server
  specified by a \code{catmaid_connection} object. If such an object is not
  specified, then one is created, by default using \code{options} of the form
  "catmaid.*".

  The connection object returned by \code{catmaid_login} will then be used
  for future requests to the CATMAID server.
}
\details{
After successful login, the \code{catmaid_connection} will contain a
  \code{cookie} field that includes a sessionid that is required for
  subsequent GET/POST operations.
}
\section{Package options}{


  You will very likely want to set some of the following package options in
  your .Rprofile file (see \code{\link{Startup}} for details)

  \itemize{

  \item{catmaid.server}

  \item{catmaid.username}

  \item{catmaid.password}

  \item{catmaid.authname}

  \item{catmaid.authpassword}

  } using code along these lines:

  \code{options(catmaid.server="https://mycatmaidserver.org/catmaidroot",
  catmaid.authname="Calvin",catmaid.authpassword="hobbes",
  catmaid.username="calvin", catmaid.password="hobbesagain")}
}
\examples{
\dontrun{
## example explicitly specifying connection options
conn = catmaid_login(server="https://mycatmaidserver.org/catmaidroot",
  authname="Calvin",authpassword="hobbes",
  username="calvin", password="hobbesagain")

## examples assuming that catmaid.* options have been set
conn = catmaid_login()
conn = catmaid_login(server='https://someotherserver.org/catmaidroot')

## now do stuff with the connection like
skel=catmaid_fetch("1/10418394/0/0/compact-skeleton", conn=conn)

## or for those who want to work at the lowest level
skel2=GET("https://mycatmaidserver.org/catmaidroot/1/10418394/0/0/compact-skeleton",
  config=conn$config)
}
}
\seealso{
\code{\link{options}}, \code{\link{Startup}}
}

