% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_connection.R
\name{catmaid_login}
\alias{catmaid_connection}
\alias{catmaid_login}
\title{Connect/authenticate to a CATMAID server, returning a connection object}
\usage{
catmaid_login(conn = NULL, ..., Cache = TRUE, Force = FALSE)

catmaid_connection(server, username = NULL, password = NULL,
  authname = NULL, authpassword = NULL, token = NULL,
  authtype = getOption("catmaid.authtype", default = "basic"))
}
\arguments{
\item{conn}{An optional \code{catmaid_connection} connection object.}

\item{...}{Additional arguments passed to catmaid_connection}

\item{Cache}{Whether to cache open connections at login so that they can be 
reused automatically.}

\item{Force}{Whether to force a new login to the CATMAID server (default 
\code{FALSE})}

\item{server}{url of CATMAID server}

\item{username, password}{Your CATMAID username and password.}

\item{authname, authpassword}{The http username/password that optionally
secures the CATMAID server.}

\item{token}{An API token (A modern alternative to provifing your CATMAID username and password). See \bold{Token based authentication} for details.}

\item{authtype}{The http authentication scheme, see
\code{\link[httr]{authenticate}} for details.}
}
\value{
a \code{catmaid_connection} object that can be used to make 
  authenticated http requests to a CATMAID server, specifically by making use
  of the \code{$config} field.
}
\description{
\code{catmaid_login} allows you to login to a CATMAID server 
  specified by a \code{catmaid_connection} object. If such an object is not 
  specified, then the last succesful connection in this R session is reused 
  if possible otherwise a new connection object is created using 
  \code{options} of the form "catmaid.*" (see details).
  
  The connection object returned by \code{catmaid_login} (or cached when 
  \code{Cache=TRUE}, the default) can then be used for future requests to the
  CATMAID server by get/query/fetch functions.
}
\details{
After successful login, the \code{catmaid_connection} object will 
  contain a \code{cookie} field that includes a sessionid that is required 
  for subsequent GET/POST operations.  When \code{Cache=TRUE} (the default) 
  the open connection object is cached and will be used when EITHER 
  catmaid_login is called with enough information to indicate that the same 
  server is desired OR (when no information about the server is passed to 
  catmaid_login) the last opened connection will be used.
}
\section{Token based authentication}{
 CATMAID now offers token based 
  authentication as an alternative to specifying you CATMAID username and 
  password. See \url{http://catmaid.github.io/dev/api.html#api-token} for how
  to get an API token. You can then set the catmaid.token package option. 
  Note that you must \bold{NOT} reveal this token e.g. by checking it into a 
  version control script.
}

\section{Package options}{

  
  You will very likely want to set some of the following package options in 
  your .Rprofile file (see \code{\link{Startup}} for details)
  
  \itemize{
  
  \item{catmaid.server}
  
  \item{catmaid.username}
  
  \item{catmaid.password}
  
  \item{catmaid.token} An alternative to using catmaid.username/password
  
  \item{catmaid.authname}
  
  \item{catmaid.authpassword}
  
  } using code along these lines:
  
  \code{options(catmaid.server="https://mycatmaidserver.org/catmaidroot", 
  catmaid.authname="Calvin",catmaid.authpassword="hobbes", 
  catmaid.username="calvin", catmaid.password="hobbesagain")}
  
  # or if you are using an API token # see 
  http://catmaid.github.io/dev/api.html#api-token 
  \code{options(catmaid.server="https://mycatmaidserver.org/catmaidroot", 
  catmaid.authname="Calvin",catmaid.authpassword="hobbes", catmaid.token = 
  "9944b09199c62bcf9418ad846dd0e4bbdfc6ee4b")}
}
\examples{
\dontrun{
## example explicitly specifying connection options
conn = catmaid_login(server="https://mycatmaidserver.org/catmaidroot",
  authname="Calvin",authpassword="hobbes", 
  username="calvin", password="hobbesagain")

## examples assuming that catmaid.* options have been set
conn = catmaid_login()
conn = catmaid_login(server='https://someotherserver.org/catmaidroot')

## now do stuff with the connection like
skel=catmaid_fetch("1/10418394/0/0/compact-skeleton", conn=conn)

## or for those who want to work at the lowest level
skel2=GET("https://mycatmaidserver.org/catmaidroot/1/10418394/0/0/compact-skeleton",
  config=conn$config)
}
}
\seealso{
\code{\link{options}}, \code{\link{Startup}}
}

