% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_annotations.R
\name{catmaid_meta_annotations}
\alias{catmaid_meta_annotations}
\alias{catmaid_get_meta_annotations}
\alias{catmaid_query_meta_annotations}
\alias{catmaid_set_meta_annotations}
\alias{catmaid_remove_meta_annotations}
\title{Get, query, set and remove CATMAID meta-annotations for annotations}
\usage{
catmaid_get_meta_annotations(annotations, pid = 1, conn = NULL, ...)

catmaid_query_meta_annotations(meta_annotations,
  with_annotations = FALSE, pid = 1, conn = NULL, ...)

catmaid_set_meta_annotations(meta_annotations, annotations, pid = 1,
  conn = NULL, ...)

catmaid_remove_meta_annotations(annotations, meta_annotations,
  force = FALSE, pid = 1, conn = NULL, ...)
}
\arguments{
\item{annotations}{annotation ids designating which annotations to
meta-annotate. IDs can be found by calling
\code{catmaid_get_annotationlist}. If a character string is given, then IDs
will be found by calling \code{catmaid_get_annotationlist}.}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}

\item{meta_annotations}{meta-annotation to add to query. Either a vector of
IDs or a character sting of meta-annotations can be given.}

\item{with_annotations}{whether or not to return the other meta-annotations
of an annotation, when using \code{catmaid_query_meta_annotations}.}

\item{force}{Whether to force the catmaid server to remove multiple
annotations (default \code{FALSE}, to provide some protection against
accidents).}
}
\description{
\code{catmaid_get_meta_annotations} Query by annotations to find
  the meta-annotations that label them.

\code{catmaid_query_meta_annotations} Query by meta-annotations
  to find the annotations that they label.

\code{catmaid_set_meta_annotations} Meta-annotate a group of
  CATMAID annotations

\code{catmaid_remove_meta_annotations} Remove meta-annotations
  from annotations.
}
\examples{
\dontrun{
## Against FAFB CATMAID server
catmaid_query_meta_annotations("ItoLee_Lineage")

# note that this is similar to:
catmaid_query_by_annotation("^ItoLee_Lineage$", type = 'annotation')
}
}
\seealso{
\code{\link{catmaid_query_by_annotation}} which has some overlap in
  functionality, \code{\link{catmaid_get_annotations_for_skeletons}},
  \code{\link{catmaid_skids}}, \code{\link{catmaid_get_annotationlist}}
}
