% Generated by roxygen2 (4.0.2): do not edit by hand
\name{catmaid_query_by_neuronname}
\alias{catmaid_query_by_neuronname}
\title{Get list of neurons/annotations querying by neuron/annotation name}
\usage{
catmaid_query_by_neuronname(query, pid = 1, maxresults = 500, raw = FALSE,
  ...)
}
\arguments{
\item{pid}{project and skeleton ids}

\item{maxresults}{The maximum number of results to return}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
a data.frame containing the results with an attribute "annotations"
  containing the annotations as a raw list
}
\description{
Get list of neurons/annotations querying by neuron/annotation name
}
\examples{
\dontrun{
catmaid_query_by_neuronname("ORN")
}
}

