% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{catmaid_query_connected}
\alias{catmaid_query_connected}
\title{Find neurons connected to a starting neuron}
\usage{
catmaid_query_connected(skid, minimum_synapses = 1, pid = 1, raw = FALSE,
  ...)
}
\arguments{
\item{skid}{skeleton ids}

\item{minimum_synapses}{Must be at least this number of synapses between
starter neuron and returned partners}

\item{pid}{project id (default 1)}

\item{raw}{Whether to return completely unprocessed data (when \code{TRUE})
or to convert the nodes and connectors lists into processed data.frames
(when \code{FALSE}, the default)}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function.}
}
\value{
A list containing two data.frames (incoming, outgoing) when
  \code{raw=FALSE}
}
\description{
Find neurons connected to a starting neuron
}
\examples{
\dontrun{
orn13a=catmaid_query_by_neuronname("13a ORN left")$skeleton_ids
catmaid_query_connected(orn13a)
}
}

