% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_skids.R
\name{catmaid_skids}
\alias{catmaid_skids}
\title{Find skeleton ids (skids) for various inputs including textual queries}
\usage{
catmaid_skids(x, several.ok = TRUE, conn = NULL, ...)
}
\arguments{
\item{x}{one or more skids or a query expression (see details)}

\item{several.ok}{Logical indicating whether we can allow multiple skids.}

\item{conn}{A \code{catmaid_connection} objection returned by 
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection 
object will be generated using the values of the \bold{catmaid.*} package 
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{additional parameters passed to \code{catmaid_query_by_annotation}}
}
\value{
\code{integer} vector of skids (of length 0 on failure).
}
\description{
An efficient way to find neuron skeleton ids interactively or in any function
that can take skids as an input.
}
\details{
If the inputs are numeric or have length > 1 they are assumed
  already to be skids and are simply converted to integers. If the input is a
  string starting with "name:" or "annotation:" they are used for a query by
  \code{\link{catmaid_query_by_name}} or
  \code{\link{catmaid_query_by_annotation}}, respectively.
}
\examples{
\dontrun{
# these are just passed through
catmaid_skids(1:10)

# nb these are all regex matches
catmaid_skids("name:ORN")
catmaid_skids("name:PN")
# there will be multiple annotations that match this
catmaid_skids("annotation:ORN")
# but only one that matches this (see regex for details)
catmaid_skids("annotation:^ORN$")
}
}
