% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_stats.R
\name{catmaid_user_history}
\alias{catmaid_user_history}
\title{Fetch user contribution history}
\usage{
catmaid_user_history(from, to = Sys.Date(), pid = 1L, conn = NULL,
  ...)
}
\arguments{
\item{from}{Starting date for history}

\item{to}{End date for history (defaults to today's date)}

\item{pid}{Project id (default 1)}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\value{
A data.frame with columns \itemize{
  
  \item full_name
  
  \item login
  
  \item id
  
  \item new_cable (in nm)
  
  \item new_connectors
  
  \item new_reviewed_nodes
  
  \item date
  
  }
}
\description{
Fetch user contribution history
}
\examples{
\dontrun{
catmaid_user_history(from="2016-01-01")
# last 2 days
catmaid_user_history(from = Sys.Date()-2)
}
}
