% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_connection.R
\name{catmaid_version}
\alias{catmaid_version}
\title{Return the CATMAID version running on the server}
\usage{
catmaid_version(conn = NULL, cached = TRUE, numeric = FALSE, ...)
}
\arguments{
\item{conn}{A \code{catmaid_connection} object. The default value of NULL 
implies that the most recent cached open connection will be used.}

\item{cached}{Whether to use the cached version number for this connection 
(see details)}

\item{...}{Additional arguments passed to \code{\link{catmaid_fetch}}}
}
\value{
A character vector containing the version
}
\description{
Return the CATMAID version running on the server
}
\details{
By default the version number for the current 
  \code{\link{catmaid_connection}} is stored on the first request after a new
  login and then the cached version number is reused. Setting 
  \code{cached=FALSE} will always force a request to the server. CATMAID 
  versions now appear to be named as YYYY.MM.DD and can therefore be 
  interpreted as a date or a tripartite version number (see examples).
}
\examples{
\dontrun{
# example of checking if server version is newer than required version
current_v=numeric_version(catmaid_version())
current_v
required_v=numeric_version("2016.1.1")
current_v >= required_v
}
}

