% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catmaid_nat.R
\name{plot3d.catmaidneuron}
\alias{plot3d.catmaidneuron}
\title{Plot skeleton and connectors for neuron retrieved from CATMAID}
\usage{
\method{plot3d}{catmaidneuron}(x, WithConnectors = TRUE, WithNodes = FALSE,
  ...)
}
\arguments{
\item{x}{A neuron to plot}

\item{WithConnectors}{logical indicating whether or not to plot connectors}

\item{WithNodes}{logical indicating whether to plot branch/end points
(default to \code{FALSE} since they will obscure the synapses).}

\item{...}{additional arguments passed to \code{\link[nat]{plot3d.neuron}}
(synapses). Default: \code{TRUE}.}
}
\description{
Plot skeleton and connectors for neuron retrieved from CATMAID
}
\examples{
\dontrun{
nl=read.neurons.catmaid(c(10418394,4453485))
plot3d(nl)
}
}
\seealso{
\code{\link[nat]{plot3d.neuron}}
}

