% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_nat.R
\name{plot3d.catmaidneuron}
\alias{plot3d.catmaidneuron}
\alias{plot3d}
\title{Plot skeleton and connectors for neuron retrieved from CATMAID}
\usage{
\method{plot3d}{catmaidneuron}(x, WithConnectors = FALSE, WithNodes = FALSE,
  soma = FALSE, ...)
}
\arguments{
\item{x}{A neuron to plot}

\item{WithConnectors}{logical indicating whether or not to plot connectors
(defaults to \code{FALSE}).}

\item{WithNodes}{logical indicating whether to plot branch/end points 
(defaults to \code{FALSE} since they will obscure the synapses).}

\item{soma}{Whether to plot a sphere at neuron's origin representing the
soma. Either a logical value or a numeric indicating the radius (default
\code{FALSE}). When \code{soma=TRUE} the radius is hard coded to 2.}

\item{...}{additional arguments passed to \code{\link[nat]{plot3d.neuron}} 
(synapses). Default: \code{TRUE}.}
}
\description{
Plot skeleton and connectors for neuron retrieved from CATMAID
}
\examples{
\dontrun{
nl=read.neurons.catmaid(c(10418394,4453485))
plot3d(nl)

# now with connectors (i.e. synapses)
plot3d(nl, WithConnectors=TRUE)
}
}
\seealso{
\code{\link[nat]{plot3d.neuron}}
}
