% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.neuron.catmaid}
\alias{read.neuron.catmaid}
\alias{read.neurons.catmaid}
\title{Read one more neurons in the NeuroAnatomy Toolbox (nat) format}
\usage{
read.neuron.catmaid(skid, pid, conn = NULL, ...)

read.neurons.catmaid(skids, pid, conn = NULL, ...)
}
\arguments{
\item{skid,skids}{One or more skeleton ids}

\item{pid}{Project id}

\item{conn}{A \code{catmaid_connection} objection returned by
\code{\link{catmaid_login}}. If \code{NULL} (the default) a new connection
object will be generated using the values of the \bold{catmaid.*} package
options as described in the help for \code{\link{catmaid_login}}.}

\item{...}{Additional arguments passed to the catmaid_GET function}
}
\description{
\code{read.neuron.catmaid} reads a single neuron, while
\code{read.neurons.catmaid} generates a neuronlist object including some
metadata information.
}
\details{
These functions provide a bridge between CATMAID and the
  neuronanatomy toolbox R package (\url{https://github.com/jefferis/nat}),
  which provides extensive functionality for analysing and plotting neurons
  within the context of temaplate brains.
}
\examples{
\dontrun{
library(nat)
nl=read.neurons.catmaid(c(10418394,4453485), pid=1)
plot3d(nl)
}
}
\seealso{
\code{\link{catmaid_GET}}
}

