% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_tags.R
\name{soma}
\alias{soma}
\alias{soma.neuronlist}
\alias{soma.neuron}
\alias{somaindex}
\alias{somaindex.neuronlist}
\alias{somaindex.neuron}
\alias{somaid}
\alias{somaid.neuronlist}
\alias{somaid.neuron}
\title{Retrieve soma information for skeletons (specified via CATMAID tags)}
\usage{
soma(x, ...)

\method{soma}{neuronlist}(x, ...)

\method{soma}{neuron}(x, ...)

somaindex(x, ...)

\method{somaindex}{neuronlist}(x, ...)

\method{somaindex}{neuron}(x, ...)

somaid(x, ...)

\method{somaid}{neuronlist}(x, ...)

\method{somaid}{neuron}(x, ...)
}
\arguments{
\item{x}{a \code{\link{neuron}} or \code{\link{neuronlist}} object}

\item{...}{additional arguments passed to methods, i.e. \code{\link{nlapply}}}
}
\value{
An Nx3 matrix (\code{soma}) or an integer vector (\code{somaid},
  \code{somaindex}) with missing values filled with \code{NA}. Should a
  neuron have multiple soma tags then \code{somaid} and \code{somaindex} will
  return both (resulting in the \code{*.neuronlist} methods returning a
  \code{list} not a \code{vector}) but \code{soma} will select the first
  coordinates (with a warning) so that it always returns just one xyz
  location.
}
\description{
\code{soma} fetches the XYZ location of the soma

\code{somaindex} fetches the raw index of the skeleton node
  associated with the soma (in range 1..N, number of vertices)

\code{somaid} fetches the identifier of the skeleton node
  associated with the soma
}
\details{
CATMAID access required. Data collected and described in cited
  publication.
}
\references{
Ohyama T, Schneider-Mizell CM, Fetter RD, Aleman JV, Franconville
  R, Rivera-Alba M, Mensh BD, Branson KM, Simpson JH, Truman JW, et al.
  (2015) A multilevel multimodal circuit enhances action selection in
  Drosophila. Nature.
}
\seealso{
\code{\link{nsoma}}
}
