% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid_metadata.R
\name{update_neuronlist}
\alias{update_neuronlist}
\title{Update a local neuronlist with new CATMAID data}
\usage{
update_neuronlist(someneuronlist, skids = names(someneuronlist),
  update = c("neurons", "names"), ...)
}
\arguments{
\item{someneuronlist}{a neuronlist object}

\item{skids}{selected skeleton IDs from \code{someneuronlist}}

\item{update}{what to update, either the selected neurons or just the names for selected neurons.}

\item{...}{additional arguments passed to \code{\link{read.neurons.catmaid}} or \code{\link{catmaid_get_neuronnames}}}
}
\description{
Use to update a large neuronlist quickly by pulling just certain neurons from CATMAID,
using \code{\link{read.neurons.catmaid}}. Option to update only the names.
}
\seealso{
\code{\link{read.neurons.catmaid}}, \code{\link{catmaid_get_neuronnames}}
}
